requirements.txt
-e git+https://github.com/chrswt/awspricingfull#egg=awspricingfull
Beaker==1.8.0
boto3==1.4.3
celery==4.0.2
click==6.6
decorator==4.0.10
Flask==0.11.1
gevent==1.1.2
greenlet==0.4.10
gunicorn==19.6.0
itsdangerous==0.24
Jinja2==2.8
ldap3==2.1.1
lxml==3.6.4
markdown2==2.3.1
MarkupSafe==0.23
nose==1.3.7
pbr==1.10.0
psycopg2==2.6.2
requests==2.11.1
Requires==0.0.3
six==1.10.0
SQLAlchemy==1.0.15
sqlalchemy-migrate==0.10.0
Tempita==0.5.2
tzlocal==1.3
uWSGI==2.0.14
webencodings==0.5
websocket==0.2.1
websocket-client==0.37.0
Werkzeug==0.11.11
